+++ Source/BRPlugins/Public/Rendering/StrokeSkeletalMeshComponent.h
