+++ Source/BRPlugins/Public/Rendering/StrokeSkeletalMeshSceneProxy.h
