+++ Source/BRPlugins/Public/Rendering/StrokeStaticMeshComponent.h
