+++ Source/BRPlugins/Public/Rendering/StrokeStaticMeshSceneProxy.h
